Add capability to store *ENTIRE* model in gdrive
It has been added the code required to mount your google drive.
This require a login steps:
1. you are given a link to visiti
2. this will redirect to a login
3. once the previous step is passed you obtain a key that proves that you have given the consensus to use a the google drive service programmatically.
4. insert back in the notebook textbox
run all the remaining cells in the notebook to let the notebook
continue autonomusly (watch out of notebook timeout . Online posts claim
that this is 90 minutes
https://help.clouderizer.com/en/articles/2240606-google-colab-faqs)
When the training is completed the notebook will save the *entire* model (not only the weights, in this way we do not have to keeep track the model to which the weights belong to)
on your google drive main directory.
Moreover the notebook implkement also a function that permit to load the models in gDrive, for later assesment. Before doing this recall to set the right name in the variable used to restore the model.
To switch from training to load I added a binary variable.
All in all this permit to run the colab training without being connected
for the entire training.

Furthermore solves #5